Update Cleaning Data with PostgresSQL Queries
Data Exploration of Covid-19 in Postgres SQL: This project is based on identifying the results of the Covid-19 across a broad spectrum of population types and also how the vaccine has helped reducing mortality. This data is also used within a visualization project to help see the data collected more cohesive manner.